oraculum/pyproject.toml
[project]
name = "oraculum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.28.0",
    "faiss-cpu>=1.10.0",
    "git-filter-repo>=2.47.0",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.9",
    "langchain-text-splitters>=0.3.7",
    "langchain>=0.3.21",
    "numpy>=1.26.4",
    "openai>=1.68.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "streamlit-javascript>=0.1.5",
    "streamlit>=1.43.2",
]
